+++ FacebookTransactionParser.sln
@@ -3,14 +3,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32616.157
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FacebookTransactionParser", "FacebookTransactionParser\FacebookTransactionParser.csproj", "{0B66DD4E-523C-4BDC-B48B-B5F435784707}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FacebookTransactionParser", "FacebookTransactionParser\FacebookTransactionParser.csproj", "{0B66DD4E-523C-4BDC-B48B-B5F435784707}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{6CC341D9-8E26-4BD4-A975-0B57C6792F82}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FacebookTransactionParserTests", "FacebookTransactionParserTests\FacebookTransactionParserTests.csproj", "{16C4E554-9A43-48AB-8D42-33CFE01BC950}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,10 +20,6 @@ Global
 		{0B66DD4E-523C-4BDC-B48B-B5F435784707}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0B66DD4E-523C-4BDC-B48B-B5F435784707}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0B66DD4E-523C-4BDC-B48B-B5F435784707}.Release|Any CPU.Build.0 = Release|Any CPU
-		{16C4E554-9A43-48AB-8D42-33CFE01BC950}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{16C4E554-9A43-48AB-8D42-33CFE01BC950}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{16C4E554-9A43-48AB-8D42-33CFE01BC950}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{16C4E554-9A43-48AB-8D42-33CFE01BC950}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
