--- FacebookTransactionParser/Contracts/ILogger.cs
@@ -1,9 +1,0 @@
-﻿namespace FacebookTransactionParser.Contracts
-{
-    public interface ILogger
-    {
-        void LogInfo(string message);
-
-        void LogError(string message);
-    }
-}
