--- FacebookTransactionParser/Contracts/ITransactionEntity.cs
@@ -1,11 +1,0 @@
-﻿namespace FacebookTransactionParser.Contracts
-{
-    internal interface ITransactionEntity
-    {
-        public string? Price { get; set; }
-
-        public string? Tax { get; set; }
-
-        public string? ShippingCost { get; set; }
-    }
-}
