+++ FacebookTransactionParser/Contracts/ITransactionProcessor.cs
@@ -1,0 +1,9 @@
+﻿namespace FacebookTransactionParser.Contracts
+{
+    public interface ITransactionProcessor
+    {
+        public Dictionary<string, decimal> GetStatementSummary();
+
+        public void ProcessOrderSummary();
+    }
+}
